+++ sdk/cosmos/src/responses/mod.rs
@@ -38,48 +38,49 @@ mod replace_document_response;
 mod replace_permission_response;
 mod replace_reference_attachment_response;
 mod replace_stored_procedure_response;
-pub use self::create_collection_response::CreateCollectionResponse;
-pub use self::create_database_response::CreateDatabaseResponse;
-pub use self::create_document_response::CreateDocumentResponse;
-pub use self::create_permission_response::CreatePermissionResponse;
-pub use self::create_reference_attachment_response::CreateReferenceAttachmentResponse;
-pub use self::create_slug_attachment_response::CreateSlugAttachmentResponse;
-pub use self::create_stored_procedure_response::CreateStoredProcedureResponse;
-pub use self::create_trigger_response::CreateTriggerResponse;
-pub use self::create_user_defined_function_response::CreateUserDefinedFunctionResponse;
-pub use self::create_user_response::CreateUserResponse;
-pub use self::delete_attachment_response::DeleteAttachmentResponse;
-pub use self::delete_collection_response::DeleteCollectionResponse;
-pub use self::delete_database_response::DeleteDatabaseResponse;
-pub use self::delete_document_response::DeleteDocumentResponse;
-pub use self::delete_permission_response::DeletePermissionResponse;
-pub use self::delete_stored_procedure_response::DeleteStoredProcedureResponse;
-pub use self::delete_trigger_response::DeleteTriggerResponse;
-pub use self::delete_user_defined_function_response::DeleteUserDefinedFunctionResponse;
-pub use self::delete_user_response::DeleteUserResponse;
-pub use self::execute_stored_procedure_response::ExecuteStoredProcedureResponse;
-pub use self::get_attachment_response::GetAttachmentResponse;
-pub use self::get_collection_response::GetCollectionResponse;
-pub use self::get_database_response::GetDatabaseResponse;
-pub use self::get_document_response::GetDocumentResponse;
-pub use self::get_partition_key_ranges_response::GetPartitionKeyRangesResponse;
-pub use self::get_permission_response::GetPermissionResponse;
-pub use self::list_attachments_response::ListAttachmentsResponse;
-pub use self::list_collections_response::ListCollectionsResponse;
-pub use self::list_databases_response::ListDatabasesResponse;
-pub use self::list_documents_response::{
+
+pub use create_collection_response::CreateCollectionResponse;
+pub use create_database_response::CreateDatabaseResponse;
+pub use create_document_response::CreateDocumentResponse;
+pub use create_permission_response::CreatePermissionResponse;
+pub use create_reference_attachment_response::CreateReferenceAttachmentResponse;
+pub use create_slug_attachment_response::CreateSlugAttachmentResponse;
+pub use create_stored_procedure_response::CreateStoredProcedureResponse;
+pub use create_trigger_response::CreateTriggerResponse;
+pub use create_user_defined_function_response::CreateUserDefinedFunctionResponse;
+pub use create_user_response::CreateUserResponse;
+pub use delete_attachment_response::DeleteAttachmentResponse;
+pub use delete_collection_response::DeleteCollectionResponse;
+pub use delete_database_response::DeleteDatabaseResponse;
+pub use delete_document_response::DeleteDocumentResponse;
+pub use delete_permission_response::DeletePermissionResponse;
+pub use delete_stored_procedure_response::DeleteStoredProcedureResponse;
+pub use delete_trigger_response::DeleteTriggerResponse;
+pub use delete_user_defined_function_response::DeleteUserDefinedFunctionResponse;
+pub use delete_user_response::DeleteUserResponse;
+pub use execute_stored_procedure_response::ExecuteStoredProcedureResponse;
+pub use get_attachment_response::GetAttachmentResponse;
+pub use get_collection_response::GetCollectionResponse;
+pub use get_database_response::GetDatabaseResponse;
+pub use get_document_response::GetDocumentResponse;
+pub use get_partition_key_ranges_response::GetPartitionKeyRangesResponse;
+pub use get_permission_response::GetPermissionResponse;
+pub use list_attachments_response::ListAttachmentsResponse;
+pub use list_collections_response::ListCollectionsResponse;
+pub use list_databases_response::ListDatabasesResponse;
+pub use list_documents_response::{
     ListDocumentsResponse, ListDocumentsResponseAttributes, ListDocumentsResponseEntities,
 };
-pub use self::list_permissions_response::ListPermissionsResponse;
-pub use self::list_stored_procedures_response::ListStoredProceduresResponse;
-pub use self::list_triggers_response::ListTriggersResponse;
-pub use self::list_user_defined_functions_response::ListUserDefinedFunctionsResponse;
-pub use self::list_users_response::ListUsersResponse;
-pub use self::query_documents_response::{
+pub use list_permissions_response::ListPermissionsResponse;
+pub use list_stored_procedures_response::ListStoredProceduresResponse;
+pub use list_triggers_response::ListTriggersResponse;
+pub use list_user_defined_functions_response::ListUserDefinedFunctionsResponse;
+pub use list_users_response::ListUsersResponse;
+pub use query_documents_response::{
     QueryDocumentsResponse, QueryDocumentsResponseDocuments, QueryDocumentsResponseRaw,
     QueryResponseMeta, QueryResult,
 };
-pub use self::replace_document_response::ReplaceDocumentResponse;
-pub use self::replace_permission_response::ReplacePermissionResponse;
-pub use self::replace_reference_attachment_response::ReplaceReferenceAttachmentResponse;
-pub use self::replace_stored_procedure_response::ReplaceStoredProcedureResponse;
+pub use replace_document_response::ReplaceDocumentResponse;
+pub use replace_permission_response::ReplacePermissionResponse;
+pub use replace_reference_attachment_response::ReplaceReferenceAttachmentResponse;
+pub use replace_stored_procedure_response::ReplaceStoredProcedureResponse;
